added extra test

+++ Dnd-Inventory/Dnd Inventory Logic/DomainModels/ItemModel.cs
@@ -12,7 +12,7 @@ namespace Dnd_Inventory_Logic.DomainModels
         public string Name { get; set; }
         public string Description { get; set; }
         public string Type { get; set; }
-        public float Weight { get; set; }
+        public int Weight { get; set; }
         public int Amount { get; set; }
         public int Price { get; set; }
         public int? sessionId { get; set; }
